#165 refactor : 기존 코드와 충돌 수정
